DEV: write bytes directly to output
Per Saj's feedback:

It would be more efficient to restructure the program so that it need not buffer all of the output. Let the operating system do all of the work. This is especially true for large amounts of data, as may be the case here.

The io.Reader and io.Writer interfaces in Go are quite powerful, and should form the basis of your internal I/O design.

Conversions of string → []byte (and vice versa) require a memory allocation. Allocations are sometimes unavoidable, but they generate garbage. The garbage must be collected, later. In real programs, it is usually better to avoid unnecessary allocations and alleviate pressure on the garbage collector. If you already have []byte slice, use the []byte slice as-is.

The fmt package is useful when you truly need a formatted string. In all other cases, all it does is make your program slower. (fmt can only work with run-time reflection.)